pyproject.toml
[project]
name = "proyectofinal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
    "opencv-python>=4.10",
    "numpy>=1.26",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]
